# Optimize calls into StreamSocketOuput to WriteFast further